Strongly typed the topic variable and some changes in page layout.
Up to now, for convenience, the variables being passed to methods and as elements of an array were chosen as any. This defeats any validation that typescript can offer at compile time. So now each topic has been defined as a JS object with title and desc (description). Also, the list of topics only displays the title and only on clicking the title does the description of the topic show.