Created form for notes and added notes to DB.
The form for notes is reactive and programmatic, the input fields can be added and removed and arranged in any sequence. Most importantly, code appeas excatly the way it was written with spaces, tabs and new line characters. The submission of the form writes the note successfully to the MongoDB database.